rename object_fwd.h to object.h
