+++ runtime/msgpack/adaptor/float.h
@@ -9,7 +9,7 @@
 //
 #pragma once
 
-#include "runtime/msgpack/object_fwd.h"
+#include "runtime/msgpack/object.h"
 
 namespace msgpack {
