+++ runtime/msgpack/object_visitor.h
@@ -12,7 +12,7 @@
 #include <vector>
 
 #include "common/mixin/not_copyable.h"
-#include "runtime/msgpack/object_fwd.h"
+#include "runtime/msgpack/object.h"
 
 namespace msgpack {
